Added Command to get all events
